pep8 changes

+++ thechopper/tests/test_buffered_subvolume_calculations.py
@@ -32,12 +32,12 @@ def generate_3d_regular_mesh(npts_per_dim, dmin, dmax):
 
 
     """
-    x = np.linspace(dmin, dmax, npts_per_dim+1)
-    y = np.linspace(dmin, dmax, npts_per_dim+1)
-    z = np.linspace(dmin, dmax, npts_per_dim+1)
-    delta = np.diff(x)[0]/2.
+    x = np.linspace(dmin, dmax, npts_per_dim + 1)
+    y = np.linspace(dmin, dmax, npts_per_dim + 1)
+    z = np.linspace(dmin, dmax, npts_per_dim + 1)
+    delta = np.diff(x)[0] / 2.0
     x, y, z = np.array(np.meshgrid(x[:-1], y[:-1], z[:-1]))
-    return x.flatten()+delta, y.flatten()+delta, z.flatten()+delta
+    return x.flatten() + delta, y.flatten() + delta, z.flatten() + delta
 
 
 def test1():
@@ -47,12 +47,12 @@ def test1():
     rng = np.random.RandomState(43)
 
     logrbins = np.linspace(-1, np.log10(250), 25)
-    rbins = 10**logrbins
+    rbins = 10 ** logrbins
 
-    subvol_lengths_xyz = np.array((1250, 1250, 1250)).astype('f4')
-    rmax_xyz = np.repeat(np.max(rbins), 3).astype('f4')
-    period_xyz = np.array((1500, 1500, 1500)).astype('f4')
-    xyz_mins = np.array((0, 0, 0)).astype('f4')
+    subvol_lengths_xyz = np.array((1250, 1250, 1250)).astype("f4")
+    rmax_xyz = np.repeat(np.max(rbins), 3).astype("f4")
+    period_xyz = np.array((1500, 1500, 1500)).astype("f4")
+    xyz_mins = np.array((0, 0, 0)).astype("f4")
     xyz_maxs = xyz_mins + subvol_lengths_xyz
 
     npts = int(2e4)
@@ -60,8 +60,7 @@ def test1():
     y = rng.uniform(0, period_xyz[1], npts)
     z = rng.uniform(0, period_xyz[2], npts)
 
-    _w = points_in_buffered_rectangle(
-        x, y, z, xyz_mins, xyz_maxs, rmax_xyz, period_xyz)
+    _w = points_in_buffered_rectangle(x, y, z, xyz_mins, xyz_maxs, rmax_xyz, period_xyz)
     xout, yout, zout, indx_out, in_subvol_out = _w
 
     explicit_mask = np.ones(npts).astype(bool)
@@ -106,7 +105,7 @@ def test2():
     _result = calculate_subvolume_id(x, y, z, nx, ny, nz, period)
     x2, y2, z2, ix, iy, iz, cellnum = _result
     assert np.all(cellnum >= 0)
-    assert np.all(cellnum < nx*ny*nz)
+    assert np.all(cellnum < nx * ny * nz)
     assert np.all(x == x2)
     assert np.all(y == y2)
     assert np.all(z == z2)
@@ -116,7 +115,7 @@ def test3():
     """Require that calculate_subvolume_id function wraps xyz points
     lying outside the box back into the box.
     """
-    Lbox = 1.
+    Lbox = 1.0
     npts_per_dim = 5
     x, y, z = generate_3d_regular_mesh(npts_per_dim, 0, Lbox)
     x[0] = -0.5
@@ -127,7 +126,7 @@ def test3():
     _result = calculate_subvolume_id(x, y, z, nx, ny, nz, Lbox)
     x2, y2, z2, ix, iy, iz, cellnum = _result
     assert np.all(cellnum >= 0)
-    assert np.all(cellnum < nx*ny*nz)
+    assert np.all(cellnum < nx * ny * nz)
     assert np.all(x2 >= 0)
     assert np.all(y2 >= 0)
     assert np.all(z2 >= 0)
@@ -140,7 +139,7 @@ def test4():
     """Place a single point at the center of each subvolume and ensure that
     the cellnum array returned by calculate_subvolume_id is correct.
     """
-    Lbox = 1.
+    Lbox = 1.0
     npts_per_dim = 5
     x, y, z = generate_3d_regular_mesh(npts_per_dim, 0, Lbox)
 
@@ -149,12 +148,12 @@ def test4():
     x2, y2, z2, ix, iy, iz, cellnum = _result
 
     #  Every cell gets exactly one point
-    for icell in range(nx*ny*nz):
+    for icell in range(nx * ny * nz):
         mask = cellnum == icell
         assert np.count_nonzero(mask) == 1
 
     #  Check a few specific cases
     mask_cellnum0 = cellnum == 0
-    assert np.all(x[mask_cellnum0] == np.array((0.1, )))
-    assert np.all(y[mask_cellnum0] == np.array((0.1, )))
-    assert np.all(z[mask_cellnum0] == np.array((0.1, )))
+    assert np.all(x[mask_cellnum0] == np.array((0.1,)))
+    assert np.all(y[mask_cellnum0] == np.array((0.1,)))
+    assert np.all(z[mask_cellnum0] == np.array((0.1,)))
